Fix various warnings and conversion issues

+++ src/reader.c
@@ -54,7 +54,8 @@ set_blank_id(SerdReader* reader, SerdNode* node, size_t buf_size)
 {
 	char*       buf    = (char*)(node + 1);
 	const char* prefix = reader->bprefix ? (const char*)reader->bprefix : "";
-	node->n_bytes = snprintf(buf, buf_size, "%sb%u", prefix, reader->next_id++);
+	node->n_bytes =
+		(size_t)snprintf(buf, buf_size, "%sb%u", prefix, reader->next_id++);
 }
 
 size_t
@@ -268,7 +269,7 @@ serd_reader_start_file(SerdReader* reader, const char* uri, bool bulk)
 		(SerdStreamCloseFunc)fclose,
 		fd,
 		name,
-		bulk ? SERD_PAGE_SIZE : 1);
+		bulk ? SERD_PAGE_SIZE : 1U);
 	serd_node_free(name);
 	return st;
 }
